gsm7/conv: add a max length when converting to GSM7.
The function sb_conv_to_gsm7 returns the number of bytes used at the end to
encode but we do not know the number of characters really encoded.

The new function, sb_conv_to_gsm7_max_len takes the max length that we can
encode and returns -1 if we try to encode more.

Refs: #22315

Change-Id: I44e8834b8e3e69750ed7ebd72624990466f056f0
Signed-off-by: Florent Tribouilloy <florent.tribouilloy@intersec.com>