el: Avoid segfault in el_show_blockers when a EV_FS_WATCH blocks.
EV_FS_WATCH was not present in typenames, leading to a segfault in
el_show_blockers when such an event blocks the event loop.

To fix that, use a function and a switch instead of a table, in order to
not forget adding corresponding cases when adding new ev_type_t in the
future.

Change-Id: I5e83b09c9adb168c3ddb54f394d5343f066818ae