mem-fifo-pool: fix potential mem-leak.
When a page get emptied, it is usually deleted except when:
* it is the current page
* there's no current "freepage"

Pages are self-contained object: they do know when they are empty and when
its worth getting removed, however that check is done only when a
deallocation occurs. As a consequence, empty pages not pointed by the
mem-fifo-pool object are definitively leaked.

This could occurs when the current page get emptied. It was then preserved
as the current page, but could be replaced by another page if it wasn't
big enough to contain the next allocation.

We fix this by ensuring the current page is never empty. When the page get
emptied, the current pointer is cleared and the page follows the standard
procedure of choosing between setting it as the freepage or deleting the
page altogether.

Change-Id: Ia212ef80ecba057c1f9286bcc49057c168d953dc
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>