ichannel: add routing hdr definition.
The new routing headers provide a way to deliver a query either to the
named entity or to a host. Each routing header contains the original
header it overload.

Named routing can target either a service or a role. Each is specialized
using an optional host identifier and an optional instance identifier.

Host routing transfer a message from a source host to a destination host.
This is the lower routing layer, higher levels get rewritten to that layer
in order to be effectively routable. The routing includes a priority in
case some queries needs to be delivered in shortest delays and a ttl in
order to avoid routing loops.

Refs: #24818
Change-Id: Ie40b9cefec8ba6e56ecd6b8a695f4074e3c4dc79
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>