ichannel: Delete "autodel" ichannels in ic_disconnect.
This patch comes to fix the asan build failures of platform
zchk-bigdata.py observed here for example:

 https://git.corp/buildbot/builders/asan/builds/403/steps/test/logs/stdio

What happened is that the platform EP was calling ic_disconnect itself on
some server peer ichannels. The platform layer is considering that if such
an ichannel is called in its on_event callback with the DISCONNECTED
event, then it will be automatically deleted by the ichannel layer, and
does NOT delete it to avoid a double-free.

The problem is that when ic_disconnect was called from outside the
ichannel layer, it was not deleting itself, and an ichannel was leaked.

Change-Id: I88b86c38b2ec71b3c77a0f19ccee85cbb371c52d