ichannel: fix big replies to be always rejected.
Since 484ef72da9c77d91aa3a9dd7b51837a8367575d6, we perform extra checks on
big messages (more than 10MB). This includes checking that replies are
expected (that is there is a slot waiting for it locally).

That code was buggy since the slot was miscomputed, keeping only the
flags. As a consequence, big replies were always rejected.

Closes: #24837
Change-Id: I625b489f60ec37b5c9d13b5733d9be4b4aefe055
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>