thr-job: fix possible starvation of worker threads.
Job stealing is done at each iteration of the worker-threads loop and if
no job is available the thread waits for a notification. However, there
were several races in that procedure.

First, we didn't make any distinction between job stealing failing because
the queues of the other threads are empty or because there was contention
with other threads. As a consequence, we could start waiting for a new job
while there were already some in the queues of the other threads. This
could potentially lead to some jobs not being consumed at all since the
waiter won't be awoken before a new job is posted.

Secondly, the thr_schedule() function used "thr_ec_signal_relaxed()" which
won't notify other threads if there is no waiters. This is racy since
another thread can be registering itself concurrently and as a consequence
it won't be awoken. That thread will never be awoken if no other job is
posted which may block forward progress of the whole process if there is
only two threads (1 or 2 CPUs available).

These issues are mostly observable when using only two threads since
having one thread blocked blocks the whole program. An heavy load is
needed since the reproduction window is very small.

Change-Id: I5b155deb10dc4c7df4d9aaaf43cf6c5c47c25763
Closes: #24782
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>