ichannel: Fix ic_disconnect following fe8c9d42f096.
fe8c9d42f096 introduced a regression in ic_disconnect: it is important to
execute the lines from "ic_choose_id(ic);" to "ic->queuable = false;" even
when is_closing is true.

This is because is_closing can be set to true by someone else than
ic_wipe, and in these cases it is important to leave the ichannel in a
state where ic_connect can be called.

This caused a segfault in the zplatform python tests on master branch
(only observed on my machine, not on the buildbots).

Change-Id: I56114fe43ea6b02b1aa4d4b73f2234bf5e0cdaa2