ASN.1: fix the use of indefinite length in members of SEQUENCE OF
When counting the elements of a SEQUENCE OF, we did not support the inner
fields with indefinite length, but we already had the right function to
support it. This patch uses the function asn1_get_field() to support that
corner case.

This patch fixes:
* Standard case (use of indefinite length in a field of a standard
  SEQUENCE OF)
* Sequence of untagged choice.

It also adds some tests for both cases.

Refs: #24976
Progress: 70%
Change-Id: I01de32dd29ce7ea09f2e690e4c932cd8eeb9286a