el: add ability to mark fd as having applicative pending events.
In case a fd has pending stuff that has already been read but is still not
processed. This patch introduces a new el_fd_mark_fired() call that will
force the event loop to process that el during it next loop. A function
that call el_fd_mark_fired() is supposed to stop its processing following
that call in order to give back the hand to other els.

Using el_fd_mark_fired() can be used to improves the fairness of a
callback that has a lot of processing to perform.

Change-Id: If2dbbffacc14ddfb09303e3ce0c70c32342e876a
Refs: #25682
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>