ic: don't consume more than 16 messages at once.
In order to improves the fairness of the ic consumption, it will not
consume more than a few messages each time it is fired by the event loop.
As a consequence, it use the newly introduced el_fd_mark_fired() when it
decides to interrupt the processing of buffered events.

Closes: #25682
Change-Id: Ie89149e2e2018c5ec1f3158f3a7a35e4e2372ee0
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>