core: support assignation to ptr in tab_for_each_ptr() iterations.
Currently, in tab_for_each_ptr(ptr, vector), the ptr object is the
iteration variable of the loop. As a consequence, if it is modified by the
loop, the behavior of the loop becomes undefined (and mostly harmful).

In order to be safer, we now have two separated variables: one for the
iteration, one for the user. The one used by the iteration is private and
must not be accessed by the user. The one used by the iteration can safely
be changed during the iteration without affecting the behavior of the
loop.

Change-Id: I2f031e757ba6e4a5cf53e7c3aad0ba63f5bb7b9c
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>