iop: Fix iop_field_is_defval for strings.
iop_field_is_defval was considering that a string does not have the
default value if it does not point on the default value, even if the
contents of the value and the default value are actually the sames.

This "optimization" was responsible for the bug #25885 observed in
Insighted:
* the core was signing a schema structure which was received by RPC and
  dumping it into a json file,
* when reloading the schema from the json file, the signature failed,
  despite the fact that both objects were considered as equal by
  iop_equals.

To fix that we have to compare the strings if the pointer equality is not
verified.

Closes: #25885
Change-Id: I83ac15dfb35777653c57a4c6ed73f8884aa6a170