el: add getter for el->fs_watch.path
when using el_fs_watch_register, if we want to use the watched directory
in the block, it must be allocated and kept on the heap.

but since the path is already duplicated on the heap and stored in
el->fs_watch.path, it can be practical to use that directly in the block.

int foo() {
    const char directory[PATH_MAX];
    ...
    watcher = el_fs_watch_register_blk(directory,
                IN_ONLYDIR | IN_MOVED_TO,
                ^void (el_t el, uint32_t mask, uint32_t cookie, lstr_t name) {
        do_stuff(el_fs_watch_get_path(el), name);
    }, NULL);
}

Change-Id: Ic4502bc553a57b8a09461789baf80aee671f7ece
Refs: #25067