iopc: remove references causing leaks
We hold too many references which can be circular, in that case nothing is
ever wiped.
For example we can very well have pkgA -> pkgB -> pkgA, so both packages
have a refcount = 2 and will never be wiped.
Same problems can happen with ifaces, structs, etc.

We actually do not need refcounting at all in most of these structures, so
just remove it completely.
After that patch, the iopc tests pass in both asan and valgrind.

Change-Id: Iaa220d0bd47ed5a70aabb53dd9cc10d6a6e29f5c