iopc: fix doxygen parsing and leaks
the following example from igloo/iop/sdk.iop would cause a leak of the
doxygen comment.

    listsGetAll  /**< private RPC for website */
        out  (MsisdnList[] mList, ImeiList[] iList, ImsiList[] imList,
              Zone[] zone)
        throw IglException;

There were two problems here, first the current parser did not expect the
back comment of the RPC here, but the front comment of the in param.
Since in param was not specified, the comment was leaked.

So this commit solves the two problems:
* support missing in or out correctly
* expect RPC back comment to be right after the RPC name, and not
  completely at end.

Here is how doxygen back comment for a RPC had to be written before this
patch:

    listsGetAll
        out  (MsisdnList[] mList, ImeiList[] iList, ImsiList[] imList,
              Zone[] zone)
        throw IglException;
    /**< private RPC for website */

Change-Id: I52e9b020491ff3cd7fe2ccd9860734fb8ef065e9