asn1-ber: allow skipping trailing field in case of indefinite length
About indefinite length
~~~~~~~~~~~~~~~~~~~~~~~
Indefinite length sequences in ASN.1 BER work quite the same as
null-terminated strings in C: we know the sequences is over because of a
special field called EOC and encoded on two bytes: 0x00, 0x00.

Most of the time, the definite length form is preferred to the indefinite
one.

Trailing fields in BER
~~~~~~~~~~~~~~~~~~~~~~
The possibility not to declare the last field(s) of a sequence is a
feature of our ASN.1 BER library.

In case of definite length, the feature worked because we simply read each
declared field then stop, we wouldn't actually read the trailing field(s)
because we already know where the sequence is ending, so we simply jump to
that position.

In case of indefinite length, we have to skip each non-declared field to
know where the sequence ends.

Change-Id: I97447eda9d6b3a2fcfc8b6bb7941984f762be62a