ic: hdr: makes routing header easily extensible.
Hardcoding the list of the routing in the ic module makes things a bit
hard to maintain and is somehow a layer violation since it brings
knowledge of how routing works to the lib-common.

In order to avoid this, put only a simple routing header in the lib-common
and make this easily extensible. THe routing header contains the base
fields (ttl, priority, tracing) which make the actual route easily
changeable without breaking the whole header.

Change-Id: I59fa75e3ca22ecb7a28d409dce2be8eed80f2a16
Refs: #24818
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>