iop-rpc-channel: set 'queuable' flag on server side when IC is connected
Problem
~~~~~~~
On client side, in the 'on_event' callback, when the event is
IC_EVT_CONNECTED, ic_is_ready() returns true and its is possible to
enqueue some queries.

On server side, in the same situation, ic_is_ready() returns false because
the flag ichannel_t::queuable is not set.

This difference may be confusing and there is no reason to prevent the
server from enqueueing some queries inside 'on_event' callback.

Solution
~~~~~~~~
- set 'queuable' flag to true in ic_mark_connected() before calling
  'on_event' callback
- to be symetric, set this flag to false in ic_disconnect() before calling
  'on_event' callback

+ coding rules

Change-Id: I077be7ffbe7499292c074c216d78ffcbb5256a06
Refs: #26507
Signed-off-by: Vianney Bajart <vianney.bajart@intersec.com>