qvector: rework qv_extend() so it can extend a vector with IOP arrays
qv_append() adds the content of a vector B into a vector A.

Before:
~~~~~~~
The type of vector A and B had to be compatible.

Now:
~~~~
Only the attributes tab and len of the "vector" B have to be compatible
with the respective attributes of vector A, allowing us to use an IOP
array or any similar structure.

Purpose:
~~~~~~~~
Simplify some code. Example taken from mcg:

 static void filter_instances_dup(qv_t(iop_instance) *vec)
 {
-    p_copy(qv_growlen(iop_instance, vec, www_g.instances->instances.len),
-           www_g.instances->instances.tab,
-           www_g.instances->instances.len);
+    qv_extend(iop_instance, vec, &www_g.instances->instances);
 }

Change-Id: Ia43882569253550615d7a454b2fb5d95ee9a53f6