module: allow module to call their methods during initialization.
Problem
-------
Let's say we have:
                       core
                        |
                      schema
                      /   \
                    mod1  mod2

And the modules are initialized like that:
          MODULE_REQUIRE(core);

During its initalization the schema is trying to call one of its method.

The states of the modules will be:
core :   REGISTERED
schema : REGISTERED
mod1:    AUTO_REQ
mod2:    AUTO_REQ

so no method will be called since there is no MAN_REQ modules.

We want to allow a module to call its method during its initialization.

Solution
--------
In a second pass run the method on all modules that have been auto
required and for which the method has not yet been called.

Change-Id: Ic8f3d5fedadaafdb0a54d2c2bbfa0b30bc32ed94