core-mem: Keep the current page when its last block is freed
Previously, when the last block of the current page was freed, the now
empty current page was discarded if there was already a free page. There
was no need to do so, as the only risk was that if we tried to allocate
more than the page size, the empty page would stay allocated forever. The
additional check in mfp_alloc corrects this.

Change-Id: I2fc5a20ea980fb1beef26c69ed0da0f577f91581