core-mem-stack: Always save block in current frame.
In the case of stack_push, the used block was not always owned by the
previous frame. This created a lot of slow paths in sp_reserve : it had to
fetch the block that had just been forgotten by stack_pop.
Now, any new block is assigned to the current frame. This reduces
branching and slow path probability in sp_reserve.

Warning : the base block must be reset in pool_wipe, all the blocks being
destroyed by the loop.

Warning : the `last' pointer in the frame must be updated too. If not,
realloc hits undefined behaviour when comparing (frame->last + asked <=
frame_end(frame)) where frame->last lies in a former block.

Change-Id: I7efbb944d4e0abab6e200c2679c34b10a5dac7dd