module: fix wrong attribute.
Since 9da425e99e, module_add_dep can take NULL for mod.
The __attr_nonnull__((1)) was not removed.

When the optimization are turned on, GCC since it knows that the argument
will not be NULL, may totally remove the code that exist in case
the argument is NULL.

so if you have:

if (!argument) {
   ...
}

will be removed when compiling with optimization, but won't if you are
compiling in debug.

Change-Id: I5b120b2094c840156750a36b9e0730a994980095