core-mem-fifo: pack structs
Modify and reorder structs so that the most frequently accessed attributes
are on the first cache line. Remove the useless start attribute from
mem_page_t. The only lost information is the total number of blocks in the
pool, which is hence not logged anymore when the pool is destroyed before
all the blocks are freed. Only the number of pages is displayed.

In mem_pool_t, the modification of the type of min_alignment allows to
reduce the size of the struct by 8 bytes. min_alignement is always less
than 16 so this shouldn't be a problem. In turn, this allows to keep all
the most used attributes of mem_fifo_pool_t (up to alive) in the first
cache line. This gives a 4-5% performance boost in the bench.

In mem_page_t, start was an alias to area and was never used.

Change-Id: Icd53fb3ea6f12133809713efa92159de66cce743