core-mem-fifo: Improve page management
In mfp_free, we don't reset the page as soon as it is empty if it is the
current page. This gives a better behaviour in situations where there is
only one block allocated at a time: instead of resetting the page each
time a block is allocated, we keep going forward on the page, and the need
for resetting is checked in mfp_alloc. In mfp_free, we also keep the page
to be deleted as freepage if it is larger than the current freepage.

We also create a new page as soon as the pool is created. This allows to
remove the check for current in mfp_alloc because current is never NULL
now.

This patch gives a ~10% performance boost in the benchs.

Change-Id: Id2625350cf339f91e1124073403bd9ca800123dd