core-mem-fifo: Use p_new instead of mmap
Replacing mmap/munmap by p_new/p_delete gives a ~10% performance boost in
the benchs:

Here is the benchmark data:
page_size   p_new    mmap    delta
     4096  0.05850  0.08593  31.92%
     8192  0.07407  0.09754  24.07%
    16384  0.09733  0.11952  18.56%
    32768  0.13477  0.15735  14.35%
    65536  0.18983  0.21294  10.85%
   131072  0.25961  0.28856  10.03%
   262144  0.37201  0.40032   7.07%
   524288  0.53528  0.55306   3.22%
  1048576  0.75730  0.76380   0.85%
  2097152  0.77117  0.88887  13.24%
  4194304  1.13518  1.25524   9.57%

The timings are the output of ./ztst-mem-bench -f, averaged over 500 runs.
The size of the memory requests was increased with the page size to keep
the number of pages allocated during the bench reasonable, without
exceeding the test system's amount of physical memory.

The default page size in the allocator is 65536.

Change-Id: Ic1f52655b04e6cf411d2db073ffec35272f1e478