core-mem-stack: Change empty value for last.
The code used NULL to flag non-existing last allocation. This needed to
check for non-NULL pointer in sp_realloc.

Now, when we need an invalid last value, we use the current allocator
position. It is guaranteed not to have been allocated yet. This value is
known to be non-NULL. This removes a check in sp_realloc fast path.

Change-Id: I01cb1107ecf9da9ffdf564036ced0a60c2f3a98d