core-mem-stack: Be more conservative with frame allocations.
This patch aims to reduce the likelihood of frame allocations by keeping
re-usable frames of reasonable size at wipe time. This requires to split
the logic between full wipe (destruction of pool) and partial wipe
(periodic cleanup of pool).
The periodic cleanup now keeps the largest block in the range 64-256 times
the mean allocated size.

This reduces the proportion of slow pathes in sp_reserve to almost-zero.

Change-Id: Ifd0a0900b7b4450d6b314f63c21069f8b714a3e3