lib-common: Replace printf by a logger in mem-bench.
We used to printf all the human-readable statistics. Now we use the logger
structures to do it properly.
When the user of a mem_bench object cannot ensure its termination before
the shutdown of the log module, he must call mem_bench_leak to register a
proper cleanup. A mem_bench module is used internally to ensure termination
of the allocated loggers.

When MEM_BENCH is active, the stack allocator requires that no t_push
happened after wiping the pool. This allows a saner behaviour.

Change-Id: Iec881b5d0ee76d198117779f3cb53defa79401f1