core-mem-fifo: Additional page reusing mechanism
When we have a newly emptied page and we already have a freepage, if the
current page is almost full, it is more interesting to put the empty page
as current than to throw it away. It potentially wastes 1/8th of the
memory of the current page, but it allows to keep the freepage longer and
sometimes avoids to allocate a new one.
This was done to tackle a bad behaviour spotted in igloo.

Change-Id: I9e3529fd9172da50ebae5f967801070747be5be6