core-mem: redefine PAGE_SIZE macro to avoid syscalls
The PAGE_SIZE macro used in core-mem-stack and core-mem-ring used to hide
a call to sysconf, which gets repeated inside the ROUND_UP macro.
Since this value is not supposed to change soon, PAGE_SIZE is redefined
explicitly to its constant value in core-mem.h

A sanity check is done in core-mem.c constructor.

Change-Id: I32350c39eac2249e606d12e9608fcb87b5e6bd96