iop: factorize references of SOAP documentation
Problem
~~~~~~~
The references of SOAP documentation are not present in the iop-soap.adoc.
They are duplicated in each product documentation. So, if we need to
add/modify/delete references, we have do it for each product.

Moroever, some of them (LMS, MCMS) are linked to the wrong version of
SOAP documentation: we support SOAP 1.1 and not SOAP 1.2.

Solution
~~~~~~~
Put references in a separate file iop-soap-references.adoc in order to be
able to insert them where we want the product documenation.

TODO mmsx: include this file in bibliography section

Change-Id: I678fa76804b907d4fb36b712e32b3cc4d32387e5
Signed-off-by: Vianney Bajart <vianney.bajart@intersec.com>