module: have a helper to retrieve a module by name.
The module code tends to hide the actual name of the module variable, but
in some cases (mostly in platform), we need get the pointer to the module
which used to break the module abstraction.

We now expose a MODULE() helper that returns a pointer to the module of
that name. This improves the consistency of the code.

Change-Id: I408523bbf0cc7a8fdedd4bd65a1bd6188f7c4d53
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>