Build: add ability to run test with testem
About testem "Unit testing in Javascript can be tedious and painful, but
Testem makes it so easy that you will actually want to write tests."

To run testem in ci mode, a configuration file is required. His name is
testem.json. So, if an testem conf file is included in ZFile it will be
handled by  the "make check" command.

Note that I did not find another solution to launch testem without
to go in the configuration file directory.

Change-Id: Icad9d6fbc94ab7ba3d279c908be7680f02897852
Refs: #26534