ichannel: writes data as soon as possible.
When processing the pending events of an ichannel, we used to process all
the incoming messages then write everything. This works perfectly well
when the processing of the incoming messages is very fast. However, when
there are a lot of pending messages, delaying the emission of the replies
cause a lot of latency because of the actual delay caused by waiting for
the ichannel to have finished processing the incoming messages (which
could be a huge amount of time), but also because we don't fully benefit
from the kernel buffering of the socket: we used to write on the socket
only a long time after the buffer has been consumed, lower the overall
efficiency of the ichannel.

We now try to write the reply immediately after emitting it, dramatically
improving the reactivity of the ichannels under heavy load.

Change-Id: I3c86b68ddfcf01cae182252fa13f4e2af836e47c
Closes: #25682
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>