qvector: rename misleading "len" parameters as "size" parameters
In qvectors semantics, the length is the number of elements in the vector
while the size is the number of elements that can fit in the chunk of
memory currently used by the vector.

The function-like macros qv_inita(), mp_qv_init(), t_qv_init() and
r_qv_init() had a parameter "len" given but the attribute "len" of the
vector was still zero after the call because the parameter only defined
the size of the vector.

This patch renames the "len" parameters as "size" when appropriate.

Also fixes '\' alignments homogeneity around the code.

Change-Id: Iad4a10d806be4f4761638cc45e7f6e6d39228c25