lib-common: add htlist_last_entry helpers + unit tests
As a htnode_t is a structure only containing the field "next", tail is a
pointer to the last htnode_t of a htlist_t. Thus, it allows us to have a
helper to get the last element of a htlist.

Clearly, it seems that the htlist architecture could be changed, there is
no need to have tail as a htnode_t **, it can be a htnode_t *. Will
probably be done later.

Change-Id: I7e3bd90d03e401a240f631d55197d85e3e0ddc5b
Signed-off-by: Ludovic Demeyer <ludovic.demeyer@intersec.com>