lib-common: htlist rework
Until now, the "tail" field of a htlist was a pointer to the "next" field
of the last element of a list. In case of empty list, "tail" was a pointer
to the "head" field of the list.

This leads to a lot of misunderstanding by developers and overcomplicate
the use of htlist for no gain.

This patch change htlist in order to have the "tail" field as a htnode_t
*, just like the "head" one. It's now pointing directly to the last
element of a list, or NULL if list empty.

Tests will follow when I7e3bd90d03e401a240f631d55197d85e3e0ddc5b will be
merged in master.

Change-Id: I6a8a798ba321976a84134cc107b5e63773ac77bc
Signed-off-by: Ludovic Demeyer <ludovic.demeyer@intersec.com>