ichannel: fix possible leak of hdr on IC_CB_DYNAMIC_PROXY
In case of __mw_dynproxy get_ic function, pxy_hdr could reference the local
scope t_stack allocated hdr and pxy_hdr is used outside of this scope.

Declare a t_scope and an hdr for the whole ic_read_process_query(), so that
the data referenced in pxy_hdr is still valid when building the new
message.

Refs: #28038
Change-Id: If98ca8822a5e47f4e0386f4f5b960b6953d8378f