ichannel: fix bad value when proxify with a pxy_hdr
In ic_read_process_query() when we proxify with the pxy_hdr then the
original hdr if present should be skipped from the data otherwise the
receiver would trigger an iop error when unpacking the value.

So we should support header replacement with dynamic proxyfication.

Add packed_hdr_len out argument to t_get_hdr_value_of_query() in order to
know the len of the original hdr inside the received data when this data is
packed.

Closes: #28038
Change-Id: Idf04f341092b69d022f8b71abdcaf6f9be1a8fb3