str-l: Check for negative values in lstr_to_uint64.
strtoull accept negative values with triggering any error, which is not
what we want in most of cases when parsing unsigned numbers.
It seems that there is no silution but to manually check the presence of a
minus sign...

Do not directly fix memtoullp helper, because it is used in many places,
and this could break some things.
lstr_to_uint64 is a recent helper, and this change seems safe.

Change-Id: If3d706abb5f92fbae9d63d71033bc6e2d92b0974