ichannel: fix leak when pre_hook fails before proxification
An ic_msg_t is allocated for proxification, this msg was allocated before
pre_hook code, when the pre_hook code failed the allocated msg was not
deleted.

Do the pre_hook before allocating the msg, this msg is not needed for the
pre_hook and as this no need to delete in case of error.

Change-Id: I19096b5ad116ce815b7781c632bde6fa4c2997c4