file-log: add functions to disable/enable file rotations
Create the functions:

 void log_file_disable_rotation(log_file_t *file);
 int  log_file_enable_rotation(log_file_t *file);

These functions should be used when the user does not want some log lines
that are related to each other to be written in separate files. The
trade-off is that the max_size attribute is applied in a less strict way.

Context and use case
~~~~~~~~~~~~~~~~~~~~
This feature is designed because of stop-segments processing in Insighted:
we're analyzing the trajectory of users basing on cell events captured
from the network. One "stop-segment" is an aggregation of several cell
events, but the whole trajectory is still composed of several
"stop-segments".

When writing the output CSV, we're interested in having a log file
rotation system that prevent us to write huge CSV files, but we're also
interested in not splitting a trajectory in two, so we can use any output
file for testing matters with only complete trajectories inside.

Change-Id: I93affb2f4d7ffc3583257633df6b9b9ec1a28d4a