file-log: add a transaction-like mode for file writing function
The function log_fwrite_transaction() executes logging operations in a
block. If the block returns an error, then the writing position of the
current log file is reset to the position it had before the call.

This functions brings an additional response to the use case that brought
functions log_file_enable_rotation() / log_file_disable_rotation(). See
the related commit message for more details.

As explained in the documentation, this function does *not* provide a
solution for concurrent log writing.

Change-Id: I4e825cda7ef66530976137d43d9af464d272a69a