Build: fix no exit immediately when a test suite fails
Since 587f7ae4e11555de16d338c57d88242a6b9ade78, the make check on mmsx
is interrupted if a test suite raises an error.

The set option -e exit immediately if a command exits with
a non-zero status. So we remove it to control the test suite status and
run all tests.

Change-Id: I781a641f0336c2d77d5fb9ee436908857df0f61b
Refs: #26534