iopc: fix multi attributes in typedef
In iopc_apply_typedef, we apply the attributes of the typedef to the
field. In the general case, we just duplicate the attribute with
iopc_attr_dup (using refcounting) and append it to field->attrs.
However in the case of a multi attribute like @allow, we append the
arguments to the right field attr. In that case we need to properly
duplicate the argument from the typedef attribute to the field attribute,
instead of transferring it, since we can apply the same typedef to several
iop fields.

In the example, @allow(a) was correctly applied to field b, but it was
then deleted (with attr->args.len = 0) and not applied to field c.

Change-Id: Icd5dba1e21ae9267d5902c058ff624f6d04fbda6
Closes: #28292