qh/qm: get rid of the hash caching parameter.
Instead, we now have two different initializers, one "default" without
caching, one with the cache activated.

For now, the code has been adapted in order to be API compatible in order
to allow a soft migration (soft, but we'll try to make it fast).

The static initializers use an ugly trick that will only work in case the
parameter given is either "false" or "true" since it's based on the size
of the given string... some cases seem to use NULL, 0 or 1, so some code
will be broken by this change.

The mp_q?_init() family don't have the cached variant anymore and does not
preserved the backward compatibility since the API is very young, and by
consequence there are a limited number of usages.

Change-Id: I42b5f018e89b871aefd2d00eff296c5f5fb13223
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>