fix strict-aliasing issues.
We introduce a new cast helper (acast) that work around the strict
aliasing requirement for cases where we really need an incompatible cast.

acast stands for "aliasing cast".

Note that we don't activate the -fstrict-aliasing flag anyway, this is too
dangerous for no mesurable gains. This patch just ensures the code is
compliant with the strict-aliasing.

Change-Id: Id20f45da9168f9dbf8ce9ad8e0d1d0df44551535
Signed-off-by: Florent Bruneau <florent.bruneau@intersec.com>