asn1-aper: fix constrained number encoding for ranges larger than 16-bits
[THIS PATCH IS A BACKPORT]

When we encode the length of a constrained number with a range over 2^16,
(which does not happen so often because of the constraint we have most of
the time), it is specified that the length has to be encoded as a fully
constrained number. The length zero is not supposed to be used in this
case so the decoded value should be 'len - 1'.

This fixes use cases in RANAP. We did not use such integers in SABP.

Change-Id: If109f8d70e6bb37c27f442c11a5f55be8ffa68d7
Signed-off-by: El Hadi Dendene <el-hadi.dendene@intersec.com>
(cherry picked from commit aacc172b4e2be144c41c80537a71c32ef339a01a)