Get rid of first argument of t_fmt.
This first argument was a pointer on the length of the allocated string,
which was really not self-evident given the name of this argument in the
macro definition (fmt).

It was never used and always set to NULL. The funny thing is that the only
place where it was not explicitly "NULL" (that is in file-log.blk), it was
actually always set to NULL...

This patch gets rid of it, making the usages of the t_fmt macro easier.
Note that if one needs the length of the allocated string, he should use
t_lstr_fmt instead.

Also:
* remove the r_fmt macro, which is never used in our entire code base.
* rename "out" parameter of mp_fmt into lenp, to make it more explicit.

Change-Id: I3b3425e8c85b9e685800766fbc0119b1f1c862b8