lib-common/xmlr: make node_close more robust
The commit 2e0cd3e09d9d8db93caa5082dd079e43b11ec17d aimed to consider
<toto></toto> as an autoclosing node (can't figure out the use case
though), and therefore allow it to be "closed" when the xmlReaderPtr
is on <toto>.

However, the check was litteraly skipping the whole following node and
it allowed to close patterns such as:

<root>
    <child>value</child>
    <garbage>garbage</garbage>
</root>

by calling xmlr_node_close() once on "<garbage>".

This patch adds unit tests for basic cases and reworks the particular
case by checking that the node is indeed empty and can be considered
autoclosing.

Change-Id: Idf3c8647035b238645100fbd9224570abd7849bb
Signed-off-by: Navid Djelveh <navid.djelveh@intersec.com>