lib-common: bin-file library
This patch adds a library to create binary log files (as used in
platform's logd). This will be useful to factorize some code.

A binary file is basically a file containing opaque data, each record
being prefixed by its length.
The files are built so that there is the offset of the next record at each offset
multiple of a constant (by default, 1M).

Change-Id: I2b22e2c91824c7f643c5bfcbce97828fd9f6cf02
Signed-off-by: Ludovic Demeyer <ludovic.demeyer@intersec.com>