build: Add an environment variable to disable some checks in ZFile
Context:
The current behavior with "make check" is to play all the tests included
in the ZFile.

Issue:
We want to be able to play web tests separately but we don't want to remove
them from the check target either.

Fix:
Add an environment variable to disable some tests by group of the same
kind.
Add a new target which is compiling the web dependencies only and running
web tests.

Change-Id: Ia89a5cd4538b7b41429cc4148a472188029f089f
Closes: #28991
Signed-off-by: Nicolas Angevelle <nicolas.angevelle@intersec.com>