iopc: fix error reporting with pretty token
the message "%s expected, but got %s instead" could display some buggy
chars when the token was not handled in the pretty_token function.

For example for an attribute, we tried to display a non printable char
(ITOK_ATTR = 129).

Also remove the silly code that allowed up to 8 static strings returned by
pretty_token and use t_fmt instead. (This allowed to call pretty_token up
to 8 times in an error formatting function)

Finally add the missing values: attr, verbatim_c and bool

Change-Id: I3b5ce54c816f2195a26239d0ac3b27ab902534ab