module: have ability to register at fork methods for shared libs
At fork methods shouldn't be enabled by default for shared libs because
when the dso is unloaded at fork functions point to invalid addresses that
trigger a segfault when forking after unload.

But there are some cases where at fork methods are needed for shared libs,
in python for example where the imported module have a refcount at 3 so
cannot be unloaded before the end of the application.

In the use case of #28469 a fork is done after opening channels and
el_fd_at_fork() was not called so child process received events from the
related fd and closed it on connexion lost from the remote then the parent
process when using this fd failed with "No such file or directory" error.

Export module_register_at_fork() and when compiling for shared libs this
function is not a constructor, then application code can enable or not
module at fork functions.

Refs: #28469
Change-Id: Iedf3dc748717b384c834575734dcc22147ce9d32