bit-wah: Adapt API to make wah storage safer.
We discovered several bugs in QTS these last days which were due to a
missing call to wah_pad32 before writing a wah on disk: some entries were
lost.

The goal of this commit is to make it safer, by:
* renaming the "data" fields into "_data", so that this is explicit that
  this should not be accessed directly in general (and to be sure to not
  miss some direct access occurrences in our code-base). Same thing for
  "pending" and "padding", mostly for consistency.
* add the helper wah_get_data that must be used to get a wah content in
  order to write it on disk. It asserts that there is no pending data, so
  that we detect potential bugs.

Also add wah_new_from_data_lstr, because it is convenient.

Change-Id: If40a7a7f216de9dcc00596c47cdad6d3ea5d6e3b