iop-dso: recursive fixup of parent class
When loading an IOP dso with use external package, we have to do the
parent class fixup recursively. Otherwise, the fixup may be incomplete
depending on the order of exported packages.
For example if we register package pkg1 and load a plugin which export
pkg2 and pkg3, with a class A in pkg1, B in pkg2 and C in pkg3 (C
inherits from B which inherits from A):
* If the plugin export pkg2 first, everything work fine.
* If the plugin export pkg3 first, trying to get the definition of C
  won't work (ex: loading a JSON file describing a C object).

Change-Id: Ie1a02c274533a7ce6904f7044ea3b95512e61823