file-bin: fix file_bin_truncate
the file-bin test did not pass on my machine, it constructed a buggy file.
the problem is in file_bin_truncate, we are working at the fd level with
ftruncate, so we need to flush the FILE * first to keep everything in
sync.

Also use xftruncate to retry in case of temp error.

Change-Id: I05ed79685a619fb32401a34166f97d6c70ffa76e