iopc: fix min constraint operator on double
When min constraints are merged, the max one is kept.

The initial value was set to FLT_MIN, however, FLT_MIN
is not the smallest double value, but the smallest not
null positive value (1e-38).

So as a consequence, all the values lower or equals to 0
was ignored, leading to a min equals to FLT_MIN.

This issue was introduced by 7e64b3099. Thus all versions
delivered after 2014-09-30 are impacted:
* B-On: 1.0.9, 1.0.10
* Igloo: 4.0.10, 4.0.11, 4.1.0-RC2, 4.1.0-beta1
* LMS: 5.1.13, 5.1.14, 5.1.15, 5.1.16
* MCG: 6.3.18, 6.3.19, 6.4.10, 6.4.11, 6.4.12, 6.4.13
* MCMS: 5.0.16, 5.0.17

However it seems that only B-On and MCG are using IOP
constraints on double at least in master branch.

> git grep "@min([0-9]*\.[0-9]*)"
* MCG: sccore.Threshold, sccore.RetryDelay
* B-On: sdk.TariffPlan

Closes: #29901
Change-Id: I150215baeeeb72a3e80cca8deb0f558b9008ac41
Signed-off-by: Nicolas Gelot <nicolas.gelot@intersec.com>